fix UI

+++ IdScanner/SocPass.UI.Domain/Interfaces/IUserRepository.cs
@@ -5,7 +5,7 @@ namespace SocPass.UI.Domain.Interfaces
     public interface IUserRepository
     {
         Task<List<User>> GetAllUsersAsync();
-        Task<User> GetUsersByIdAsync(long? id);
+        Task<User> GetUsersByIdAsync(int? id);
         Task<User> AddUserAsync(User user);
         Task<User> UpdateUserAsync(User user);
         Task<string> DeleteUserAsync(int id);
